+++ src/components/TodoComponent/SingleTodo.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from "react";
-import { Todo } from "../model";
+import { Todo } from "../../model";
 import { AiFillDelete, AiFillEdit } from "react-icons/ai";
 import { MdDone } from "react-icons/md";
 type Props = {
